Update VMtranslator.py
The previous version was adding the Bootstrapping initial assembly code in to file even if there is only one file, which give errors and a grade of 55 in the project8 submission. 
by adding a condition so that the initial bootstrapping code will only be written if there is more then 1 file will get 100% grades. Kindly consider the pull request and thanks for providing the help.